+++ Client/src/main/java/Pojo/Hibernate/AddizionaleComunale.hbm.xml
@@ -11,18 +11,9 @@
     <property name="codice" column="codice"/>
     <property name="comune" column="comune"/>
     <property name="anno" column="anno"/>
-    <property name="aliquotaZero" column="aliquotaZero"/>
-    <property name="aliquotaUno" column="aliquotaUno"/>
-    <property name="aliquotaDue" column="aliquotaDue"/>
-    <property name="aliquotaTre" column="aliquotaTre"/>
-    <property name="aliquotaQuattro" column="aliquotaQuattro"/>
-    <property name="aliquotaCinque" column="aliquotaCinque"/>
-    <property name="fasciaZero" column="fasciaZero"/>
-    <property name="fasciaUno" column="fasciaUno"/>
-    <property name="fasciaDue" column="fasciaDue"/>
-    <property name="fasciaTre" column="fasciaTre"/>
-    <property name="fasciaQuattro" column="fasciaQuattro"/>
-    <property name="fasciaCinque" column="fasciaCinque"/>
-    <property name="sogliaEsente" column="sogliaEsente"/>
+    <property name="aliquota" column="aliquota"/>
+    <property name="redditoMinimo" column="redditoMinimo"/>
+    <property name="redditoMassimo" column="redditoMassimo"/>
+
   </class>
 </hibernate-mapping>
